Removed a compiler warning related to float and using cuda memory allocation for device memory...

+++ trunk/cuda/img_transform/blacknwhite/types.h
@@ -1,7 +1,7 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-typedef double FLOAT;
+typedef float FLOAT;
 
 typedef unsigned int UINT32;
 typedef unsigned long long int UINT64;
